implemented edit task functionality
this component implements the option to edit an already existing task. The edit button on each task will open a modal. The values are filled out in a form in this modal. Changes to the values will be saved in the tasks_list array after making modifications to the task. This automatically updates the displayed list in the tasks component